+++ source/TestCommon/documents/release-notes/release-notes.md
@@ -20,7 +20,7 @@
     - Method `InsertFromCsvFileAsync`
 
 ## Version 5.2.1
-~~~~
+
 - Extended `IntegrationTestConfiguration` with the option to pass in `DefaultAzureCredential` from the outside.
 
 ## Version 5.2.0